refactor based on comments
