+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/call/CallsScope.kt
@@ -1,8 +1,8 @@
 package com.wire.kalium.logic.feature.call
 
-import com.wire.kalium.logic.feature.call.usescase.GetOngoingCallsUseCase
-import com.wire.kalium.logic.feature.call.usescase.GetOngoingCallsUseCaseImpl
-import com.wire.kalium.logic.feature.call.usesCase.StartCallUseCase
+import com.wire.kalium.logic.feature.call.useCase.GetOngoingCallsUseCase
+import com.wire.kalium.logic.feature.call.useCase.GetOngoingCallsUseCaseImpl
+import com.wire.kalium.logic.feature.call.useCase.StartCallUseCase
 import com.wire.kalium.logic.sync.SyncManager
 
 class CallsScope(
