+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/call/useCase/StartCallUseCase.kt
@@ -1,4 +1,4 @@
-package com.wire.kalium.logic.feature.call.usesCase
+package com.wire.kalium.logic.feature.call.useCase
 
 import com.wire.kalium.calling.CallType
 import com.wire.kalium.calling.CallingConversationType
